fix: 32bit compatible

+++ common/types.h
@@ -5,9 +5,9 @@ typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
 typedef unsigned long long uint64_t;
 
-typedef uint64_t size_t;
-typedef uint64_t phys_addr_t;
-typedef uint64_t uintptr_t;
+typedef uint32_t size_t;
+typedef uint32_t phys_addr_t;
+typedef uint32_t uintptr_t;
 
 #ifndef NULL
 #define NULL ((void*)0)
